refactor: rename the package and reorganize modules
