+++ dev-env.yaml
@@ -1,9 +1,11 @@
-name: gaigar
+name: gaigar-dev
 channels:
   - conda-forge
   - bioconda
 dependencies:
+  - black=24.10.0
   - demes=0.2.3
+  - flake8=7.1.1
   - h5py=3.10.0
   - joblib=1.3.2
   - msprime=1.3.1
@@ -13,6 +15,7 @@ dependencies:
   - python=3.9.19
   - pyranges=0.0.129
   - pytest=8.1.1
+  - pytest-cov=5.0.0
   - scikit-allel=1.3.7
   - scikit-learn=1.4.1.post1
   - scipy=1.12.0
@@ -21,3 +24,4 @@ dependencies:
   - pip:
       - seriate==1.1.2
       - torch==2.2.0
+      - -e .
