--- env.yaml
@@ -1,23 +1,0 @@
-name: gaia
-channels:
-        - conda-forge
-        - bioconda
-dependencies:
-        - demes=0.2.3
-        - h5py=3.10.0
-        - joblib=1.3.2
-        - msprime=1.3.1
-        - numpy=1.26.4
-        - pandas=2.2.1
-        - pip=24.0
-        - python=3.9.19
-        - pyranges=0.0.129
-        - pytest=8.1.1
-        - scikit-allel=1.3.7
-        - scikit-learn=1.4.1.post1
-        - scipy=1.12.0
-        - tskit=0.5.6
-        - pyyaml=6.0.1
-        - pip:
-                - seriate==1.1.2
-                - torch==2.2.0
