+++ tests/generators/test_GenomicDataGenerator.py
@@ -19,8 +19,8 @@
 
 import os, pytest
 import numpy as np
-from gaia.utils import read_data, create_windows
-from gaia.utils.generators import GenomicDataGenerator
+from gaigar.utils import read_data, create_windows
+from gaigar.generators import GenomicDataGenerator
 
 
 @pytest.fixture
